tools/zep_dispatch: allow to pin nodes to MAC address
Positions in the topology are handed out first come first serve.
This makes it hard to have a reproducable setup if the ZEP dispatcher is
restarted (e.g. to modify the topology).

Fix this by allowing to specify a MAC address for each node.
If a node with a pinned MAC address connects, it will always be assigned to
the same topology node.