rust: Treat Cargo.lock files as opaque
Cargo.lock files pin the versions of all transitive dependencies, and
are left that way by Cargo unless requested manually (`cargo update`) or
necessitated by a change in dependencies (if the manually maintained
Cargo.toml says `>0.5.2` and .lock says `0.5.1`, the latter is reset).

They are fixed (as opposed to .gitignoring them and letting each user
autogenerate them) to ensure that third party changes do not break RIOT
CI builds; for updates, the changes from a `cargo update` can still be
committed and then run through CI checks. (This is analogous to how pkg
versions are pinned).

They are set to binary because their diffs are usually not practical to
read.