boards/feather-m0: sx127x params only valid for lora vairiant
In the feather-m0 board the sx127x pin parameters are hardcoded for the
feather-m0-lora variant, but this is not compatible with other
configurations.

For example in the feather-m0 adalogger PA06 and PA08 are used for the
SD card. When combined with the RFM95W FeatherWing the IRQ/CS/RST pins
must be wired, and then the sx127x pins must be configured to match the
wiring (which will be necessarily different from the lora variant
configuration).