pkg/esp32_sdk_libs: add requirements to PKG_PREPARE
The shared `build-libs` directory needs to be available for
modules/packages that depend on the SDK before that package
is eventually compiled.

Packages are downloaded, patched, prepared before any module
is compiled. By adding the directory creation as a dependency
of `PKG_PREPARE` we make sure the rule is run before compilation
starts.