sys/event/callback: in init set list_node.next to NULL
When using static initializers uninitialized fields are set to 0, or
NULL for pointers. But when using event_callback_init() on non
static variables event_callback_t structure may hold non 0 values.
This will lead to the event never being called since if super.list_node.next
is not NULL as it is considered already in the event queue and therefore
not touched.