drivers/dose: only disable watchdog when transiting from RECV state
The DOSE watchdog should only run if at least one interface is in RECV
state.

That means it must be enabled when entering RECV state and disabled
when leaving RECV state.

The watchdog was *always* disabled on a transition to IDLE.
This is wrong: If there are two interfaces and one is in RECV state
but the other did just SEND something and transitions to IDLE state
from SEND state, it would still disable the watchdog.

Fix this by only disabling the watchdog if the current state is RECV.