dist/tools/usb-serial/ttys.py: Add regex support
Use regular expressions instead of matching literal strings for all
attributes instead of the serial number to be more flexible.

The idea is to aid users when multiple boards are connected to a machine
but each of a different type. If a board would define filter arguments
for dist/tools/usb-serial/ttys.py it could be possible to detect the
right TTY automatically, without user-side modifications. However,
some boards present different model names depending on the firmware
version of the programmer. Support for regular expressions can provide
the flexibility needed to just match all revisions and variants of a
board.