rust: Bring lock files in sync with toml files
The riotmodules dependency was missed in a9be25e2, and while generally
it does not severly harm builds, it does harm builds riotdocker CI where
branches are switched, which is blocked by dirty files in the checkout.

The riot-wrappers version was missed when what is now 44fe9f51 was
rebased onto 9c4836c9.