boards/common/nrf52: fix timer config
The `channels` member should not be set to the number of hardware
channels *n*, but to *n* - 1 instead. The last channel is implicitly
used in `timer_read()`. Hence out of *n* hardware channels, only *n* - 1
are available to the application.

This fixes a bug introduced by 7db6ff2b445f10eddfc6c9c97d9767e2e294354f
which incorrectly set the channel number to *n* rather than to
*n* - 1.