CODEOWNERS: clean up
This drops people who have not contributed to RIOT (including, most
importantly, reviews) for at least a year.

For contributors it is a bit annoying to have a long list of assigned
reviewers, but none of them will ever show up to do the review. This
is especially true for new contributors who don't know which of the
reviewers are active which are ["Karteileichen"][1].

[1]: https://de.wikipedia.org/wiki/Karteileiche