doc: Allow resizing of side-nav
Currently it is not possible to resize the side-nav of the RIOT doxygen website in firefox.
This is caused by the window resize_handler being called even for resizing inside the window.
The resize_handler resets the side-nav to a cached size preventing resizing. This commit fixes
this by checking if the window was actually resized.