rust: Make dependencies visible
The Cargo.lock of rust_riotmodules_standalone should explicitly list all
versions of software that would ever be pulled in by enabling RIOT
modules backed by a Rust crate. That way, the Cargo.lock file will not
spontaneously change because the software is being built with a
different set of modules enabled.

This did not happen before, because while the optional dependencies
could all be selected on the command line, they were not visible as
possible features to `cargo update`, and were thus not encoded in the
Cargo.lock file.