boards/waspmote-pro: deprecate board
The board has history of bugs that (such as broken timer configuration)
that would directly bite any users, but still have gone unnoticed for
significant time periods due to lack of testing. It appears, that none
of the core contributors has access to the boards and the number of
users is limited.