boards/adafruit-grand-central-m4-express: provide arduino features
The board is compatible with Arduino UNO and Arduino MEGA shields and
has an ISP connector, so this adds the corresponding features.

This adds the Arduino I/O-mapping for the ISP SPI and provides the
corresponding feature.

It appears that the SPI on D11/D12/D13 cannot be provided by a SERCOM,
this is under clarification at [1]. For now, no I/O mapping for that
SPI bus is provided.

[1]: https://forums.adafruit.com/viewtopic.php?t=214093