sys/stdio: add printf_long_long
This adds the new `printf_long_long` module that can be used to enable
printing of `long long` and `unsigned long long`.

This has been implemented for `mpaland-printf`. In addition, this module
is a default module for 32-bit and 64-bit systems if `mpaland-printf` is
used, 64 bit support is not too expensive for them. (And on 64-bit
systems support for long long is needed for support of `%p`, which is
pretty basic.)

This is mainly useful for MSP430, where otherwise `mpaland-printf` would
require more memory than newlib's implementation.