examples/basic/subfolders: fix test
At least on the samr21-xpro used for hardware in the loop testing, the
terminal did not attach in time after the reset to capture all of the
output.

This adds the `test_utils_interactive_sync` module to wait for an `s`
to be received via `stdio` before starting to print to sync with the
test scripting.