boards/feather-m0: correct ADC settings
The current ADC settings set the reference to the internal 1V
reference with a x1 gain. This means that the maximum voltage
that can be measured with any analog input is 1V, which is
usually not sufficient for most applications.
Therefore the default value is changed to use the external VDDA
input, which is 3.3V (internally divided by 2) and the gain to
x1/2. This allows to measure voltages up to VDDA on any analog
input.