pkg/qdsa: drop package
This package seems to implement the [qDSA signature scheme][1], which
seemed to not have received the level of attention and scrutiny from the
crypto community one would expect before adopting it outside of lab
environments and experiments.

Hence, it might be better hosted on a repo that focuses exclusively on
research, rather than in a repo of a general purpose OS that is intended
to be used outside of lab environments as well.

[1]: https://link.springer.com/chapter/10.1007/978-3-319-70697-9_10