drivers/sx126x: fix hardware detection on init
When the CIPO line is pulled high, the driver will initialize even when
no chip is connected. This adds a read of the configured modulation,
which should be LoRa after a reset (even when configured to e.g. GFSK
prior). The packet type contains both one and zero bits, so chance
to read the only accepted byte from a floating SPI bus is relatively
low.

Co-authored-by: crasbe <crasbe@gmail.com>