tests/unittests: fix failure with CONFIG_SOCK_URLPATH_MAXLEN=128
The test checking for a failure on a too long URL path will fail when
allowing longer URL paths. E.g. in `makefiles/suit.inc.mk` we bump
`CONFIG_SOCK_URLPATH_MAXLEN`  to 128 byte. This changes the test
case to provide a path that is indeed longer than 128 byte to also
fail in that case.