doc: use https://api.riot-os.org for API doc
The documentation at https://guide.riot-os.org has become the defacto
entry point for RIOT documentation. We may dedice to let
https://doc.riot-os.org point to that doc, rather than to the API doc.

In order to keep the links in the repo working when we pull through,
this updates all doc to the API doc to `https://api.riot-os.org`. This
already works today and will keep working even if
`https://doc.riot-os.org` stops pointing to the API doc.